Improved the speed of hashing.
Combining of hashcodes of cards in battlefield must be invariant to card
order. This was done with sorting of hashcode prior to combining them. Now
it is done with XOR instead.
In Hashcalculator API the hash combining and hash caching was seperated.
All IHashable interface implementation have to be changed accordingly.