Dread's ability caused Sword of Feast and Famine to be destroyed with equiped creture.
Dread's continious effect is applied to all permanets. This is ok since
every permanent can eventualy become a creture. Triggered abilities of
swords (e.g Sword of Feast and Famine), set triggered abilities source to
the creature they are attached to. So when a combat damage was dealt the
sword was also destroyed since it looked like a creature to which it was
attached.