Eliminated state copy bugs, which caused degredation of performance in some situation. Added dection for this kind of bugs.
Performance of the game was severely degredated when some cards were
used(e.g. Sword of Fire and ice). This was due to incorect context used in
card definitions, the outer context C was used when the inner context c
was needed. As a result of this, state copy was not done properly, some
objects had references to originals and not to their copies. When such an
object was the Change tracker object a restore of game copy affected
original game, which caused the mentioned performance issues. A detection
for such bugs was implemented, and this kind of bugs were eliminated
elsewhere.