Stack overflow debug support, bugs.
Stackoverflow exception can occur if there is some infinite loop when
ai considers spells. Before every ai search a game is saved and after the
search finishes (with no exception) this save file is deleted. If a
stackoverflow occurs this file won't be deleted and the error can then be
reproduced by loading it in the debugger.

Fixed ui bug when cycling abilities were shown in ui even if controller
didn't have mana to pay them.