Modified target selector interface.
Target selector interface did not enforce the zone specification for
validators that need zone information. This caused some hard to detect
bugs. For example it was possible to specify the following:

trg.Is.Card(c => c.Is.Creature())

which is incorrect since the zone specifier is missing.

The correct would be:

trg.Is.Card(c => c.Is.Creature()).On.Battlefield

The interface was changed to enforce this at compile time.