Prevent Cycles in Mana ability activation
When checking available mana, mana source CanActivate method is called first.
This method checks if there is any additional cost to activate mana ability.
If additional cost needs mana, a mana check is performed which can again call
CanActivate method of the mana source. This could be repeated endlessly.

To prevent this a special sourceTapped variable was added to TapOwner cost.

This commit moves this functionaliy to ManaAbility class, because this
could also affect other cost types not only TapOwner.