Fix Yisan, the Wanderer Bard
Added new cost AddCounterCost, because in Yisan add counter is a cost not
an effect. Every permanent gets a small score boost, to encourage adding
counters even if this is the only thing that gets done. In Yisan's case
this allows that the Yisan gets the counters needed to fetch bigger
creatures.