Fix blocking attackers with trample
When assigning additional blockers to attackers, trample damage was not
considered. Also when players life would become less than 0, creatures
were not assigned to attackers with trample, because the score gain was
too low.

This caused that in some situations player lost the game when such loss
could be prevented.