Add support for planeswalkers
All planeswalker rules are implemented. AI for planeswalkers is
currently very primitive.

AI will always attack planeswalker with highest score and will always
redirect non combat damage to planeswalker. When you controll 2
planeswalkers with same name the one with lower score is sacrificed.

The game now supports executing non-branching decisions inline instead
only via statemachine (via Execute method). This was needed because
whether to redirect damage to planeswalker or not could not be done
inline and the whole damage system would need to be modified.

Emblems were added to describe permanent modifiers which would otherwise
not be visible in UI.